Adds support for a new --migrate-missing flag for up.
This flag allows running migrations out of order. It operates by grabbing the
  set of available migrations, and the set of already run migrations and taking
  their complement. That set of migrations is then run.

This flag is only available on `up`, it makes no sense on `down`.

**WARNING:** If you key your migrations numerically, this could be dangerous
  as it could result in migrations with the same prefix being run in an
  unspecified order. If you key by date, you should be fine. Ideally, write
  migrations that are non-destructive and non-dependent and you'll bee 100%
  fine.

Also adds a few straggling tests. Such as ensuring migrations are run in the
  right order, with or without the new flag. New test to ensure latest
  migration is always based on name.

Docs for the new feature and a warning on the arg help.